social_network/requirements.txt
Django==3.2.20
djangorestframework==3.12.4
django-filter==23.5
django-cors-headers==3.14.0
django-autoslug==1.9.9
django-templated-mail==1.1.1
djoser==2.1.0
drf-extra-fields==3.7.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.21.7

python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
oauthlib==3.2.2
whitenoise
social-auth-app-django==4.0.0
social-auth-core==4.5.3
python3-openid==3.2.0

gunicorn==20.1.0 ; sys_platform != "win32"
psycopg2-binary==2.9.9

asgiref==3.7.2
sqlparse==0.4.4
pytz==2023.3.post1

pillow==10.2.0
reportlab==4.0.8

certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
urllib3==2.2.1
PyJWT==2.8.0

coreapi==2.3.3
coreschema==0.0.4
itypes==1.2.0
uritemplate==4.1.1

Jinja2==3.1.3
MarkupSafe==2.1.5

cffi==1.16.0
pycparser==2.21
cryptography==42.0.5

coverage==7.4.4
defusedxml==0.7.1
six==1.16.0

pytest-django==4.5.2

httpx==0.28.1
django-cors-headers==3.14.0